@@ -1,7 +1,7 @@
 import { inject, Injectable } from "@angular/core";
 import { Observable, of } from "rxjs";
 import { Note } from "../types/note";
-import {HttpClient} from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { getDefinedProps } from "../helpers/common.helpers";
 import { NOTES_BASE_URL } from "./notes-service.config";
 
